fixed code smell
