+++ packages/gafl-webapp-service/src/pages/contact/digital-licence/licence-confirmation-method/route.js
@@ -7,6 +7,7 @@ import GetDataRedirect from '../../../../handlers/get-data-redirect.js'
 import { nextPage } from '../../../../routes/next-page.js'
 import { HOW_CONTACTED } from '../../../../processors/mapping-constants.js'
 import { isPhysical } from '../../../../processors/licence-type-display.js'
+import { mobilePhoneValidator } from '../../../../processors/contact-validator.js'
 
 const getData = async request => {
   const permission = await request.cache().helpers.transaction.getCurrentPermission()
@@ -24,25 +25,18 @@ const getData = async request => {
   }
 }
 
-export const mobilePhoneRegex = /^((\+44)(\s?)|(0))(7\d{3})(\s?)(\d{3})(\s?)(\d{3})$/
-export const mobilePhoneValidator = Joi.string()
-  .trim()
-  .pattern(mobilePhoneRegex)
-  .replace(mobilePhoneRegex, '$2$4$5$7$9')
-  .example('+44 7700 900088')
-
 const validator = Joi.object({
   'licence-confirmation-method': Joi.string()
     .valid('email', 'text', 'none')
     .required(),
-  email: Joi.alternatives().conditional('licence-confirmation-method', {
-    is: 'email',
-    then: validation.contact.createEmailValidator(Joi),
-    otherwise: Joi.string().empty('')
-  }),
   text: Joi.alternatives().conditional('licence-confirmation-method', {
     is: 'text',
     then: mobilePhoneValidator,
+    otherwise: Joi.string().empty('')
+  }),
+  email: Joi.alternatives().conditional('licence-confirmation-method', {
+    is: 'email',
+    then: validation.contact.createEmailValidator(Joi),
     otherwise: Joi.string().empty('')
   })
 }).options({ abortEarly: false, allowUnknown: true })
